+++ src/main/java/chess/controller/ChessGameController.java
@@ -2,6 +2,7 @@ package chess.controller;
 
 import chess.domain.BoardFactory;
 import chess.domain.ChessGame;
+import chess.domain.piece.Column;
 import chess.domain.piece.Position;
 import chess.view.InputView;
 import chess.view.OutputView;
