+++ src/test/java/chess/domain/controller/ColumnTest.java
@@ -3,7 +3,7 @@ package chess.domain.controller;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
-import chess.controller.Column;
+import chess.domain.piece.Column;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
